requirements.txt
requests~=2.31.0
python-dotenv~=1.0.0
pandas~=2.1.4
numpy~=1.26.3
matplotlib~=3.8.2
seaborn~=0.13.1
openai~=1.7.0
scikit-learn~=1.3.2
tensorflow~=2.15.0
keras~=2.15.0
streamlit~=1.30.0
flask~=3.0.0